Finalização do processo Ocpp

+++ src/main/java/br/com/gocharg/dto/ocpp/json/request/CancelReservationRequest.java
@@ -3,8 +3,10 @@ package br.com.gocharg.dto.ocpp.json.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class CancelReservationRequest {
-  private Integer reservationId;
+  private BigInteger reservationId;
 }
